added test for CronTicker.Stop
