+++ cronticker/ticker_test.go
@@ -66,5 +66,78 @@ func TestExtractTZ(t *testing.T) {
 	if expected != tz {
 		log.Fatalf("expected %q, got: %q", expected, tz)
 	}
+}
 
+func TestCronTicker_Stop(t *testing.T) {
+    ticker, _ := NewTicker("@daily")
+
+    timeoutTimer := time.NewTimer(2 * time.Second)
+
+    kCopy := ticker.k
+    ticker.Stop()
+Outer:
+    for {
+        select {
+        case <-kCopy:
+            break Outer
+        case <-timeoutTimer.C:
+            log.Fatal("Expected message on ticker 'k' channel within 2 seconds, but did not receive one")
+        }
+    }
+}
+
+// Examples for documentation
+
+func ExampleNewTicker() {
+	// The Cron schedule can be in Unix or Quartz format. Directives like
+	// '@weekly' or '@daily' can also be parsed as defined in the
+	// package github.com/robfig/cron/v3.
+
+	// You may add the TimeZone/location to the beginning of the cron schedule
+	// to change the time zone. Default is UTC.
+
+	// Example: "TZ=America/Los_Angeles 0 0 * * *"   -> Unix format: Daily at 12 AM in Los Angeles
+	// Example: "TZ=America/Los_Angeles 0 0 0 * * ?" -> Quartz format: Daily at 12 AM in Los Angeles
+	// Example: "TZ=America/Los_Angeles @daily"      -> Directive: Daily at 12 AM in Los Angeles
+	// Example: "@daily"                             -> Directive: Every day at 12 AM UTC
+
+	ticker, err := NewTicker("@daily")
+	if err != nil {
+		log.Fatal(err)
+	}
+	defer ticker.Stop()
+
+	tick := <-ticker.C
+	log.Print(tick)
+}
+
+// If you want to change the cron schedule of a ticker
+// instead of creating a new one you can reset it.
+func ExampleCronTicker_Reset() {
+	ticker, err := NewTicker("TZ=UTC 0 0 0 ? * SUN")
+	if err != nil {
+		log.Fatal(err)
+	}
+	defer ticker.Stop()
+
+	<-ticker.C
+	log.Print("It's Sunday!")
+
+	err = ticker.Reset("TZ=UTC 0 0 0 ? * WED")
+	if err != nil {
+		log.Fatal(err)
+	}
+
+	<-ticker.C
+	log.Print("It's Wednesday!")
+}
+
+func ExampleCronTicker_Stop() {
+	ticker, err := NewTicker("TZ=UTC 0 0 0 ? * SUN")
+	if err != nil {
+		log.Fatal(err)
+	}
+	defer ticker.Stop()
+
+	<-ticker.C
 }
